requirements.txt
quandl
pykafka
pandas
requests
cassandra-driver
